added log transform to input data as whitening layer
