+++ dataset_models/sdo/aia/aia.py
@@ -6,7 +6,8 @@ import psutil
 import random
 import math
 from keras.models import load_model
-
+from keras.layers import Lambda
+from keras import backend as K
 
 class AIA:
     """
@@ -159,38 +160,89 @@ class AIA:
         print "Training " + str(starting_training_count) + "-> " + str(len(self.train_files))
         print "Validation " + str(starting_validation_count) + "-> " + str(len(self.validation_files))
 
-    def get_centering_tensor(self):
+    def get_centering_tensor(self, transformation="log"):
         """
         Get a tensor for centering the data on the GPU.
+        todo: the log transformed value is for base 10, but the operation is base e.
         """
         ret = []
-        x_mean_vector = [
-            0.5015,
-            3.225,
-            111.2,
-            170.6,
-            57.03,
-            7.897,
-            1.187,
-            21.98
-        ]
+        if transformation == "log":
+            x_mean_vector = [
+                0.04378,
+                0.3213,
+                1.611,
+                1.817,
+                1.306,
+                0.5766,
+                0.1092,
+                0.8212
+            ]
+        else:
+            x_mean_vector = [
+                0.5015,
+                3.225,
+                111.2,
+                170.6,
+                57.03,
+                7.897,
+                1.187,
+                21.98
+            ]
         return np.array(x_mean_vector).reshape((1,1,1,self.input_channels))
 
-    def get_unit_deviation_tensor(self):
+    def get_unit_deviation_tensor(self, transformation="log"):
         """
         Get a tensor for changing the data to have unit variance.
+        todo: the log transformed value is for base 10, but the operation is base e.
         """
-        x_standard_deviation_vector = [
-            3.593,
-            11.11,
-            160.0,
-            246.3,
-            98.08,
-            13.45,
-            3.238,
-            24.78
-        ]
+        if transformation == "log":
+            x_standard_deviation_vector = [
+                0.1622,
+                0.386,
+                0.748,
+                0.74238,
+                0.71,
+                0.5592,
+                0.2605,
+                0.7978
+            ]
+        else:
+            x_standard_deviation_vector = [
+                3.593,
+                11.11,
+                160.0,
+                246.3,
+                98.08,
+                13.45,
+                3.238,
+                24.78
+            ]
         return np.array(x_standard_deviation_vector).reshape((1,1,1,self.input_channels))
+
+    def get_whitening_layer(self, transformation="log"):
+        """
+        Give a layer that will run on-GPU for centering and scaling the data.
+        This makes the gradient better defined.
+        """
+        centering_tensor = self.get_centering_tensor(transformation=transformation)
+        scaling_tensor = self.get_unit_deviation_tensor(transformation=transformation)
+        def output_of_lambda(input_shape):
+            return input_shape
+        def whiten(x):
+            x = K.tf.subtract(x, centering_tensor)
+            x = K.tf.divide(x, scaling_tensor)
+            return x
+        def log_whiten(x):
+            x = K.tf.log1p(x)
+            return x
+        if transformation == "log":
+            def whitener(x):
+                x = Lambda(log_whiten, output_shape=output_of_lambda)(x)
+                x = Lambda(whiten, output_shape=output_of_lambda)(x)
+                return x
+            return whitener
+        else:
+            return Lambda(whiten, output_shape=output_of_lambda)
 
     def get_x_data(self, filename, directory, image_count=2, current_data=None):
         """
