+++ network_models/xray_flux_forecast/fdl3/version3.py
@@ -174,18 +174,9 @@ x_prior = input_prior_image
 
 print "constructing network in the Keras functional API"
 
-# This tensor pulls the mean and std deviation from the dataset model
-# to whiten the inputs. This makes the gradient better defined.
-centering_tensor = aia.get_centering_tensor()
-scaling_tensor = aia.get_unit_deviation_tensor()
-def output_of_lambda(input_shape):
-    return input_shape
-def whiten(x):
-    x = K.tf.subtract(x, centering_tensor)
-    x = K.tf.divide(x, scaling_tensor)
-    return x
-x = Lambda(whiten, output_shape=output_of_lambda)(x)
-x_prior = Lambda(whiten, output_shape=output_of_lambda)(x_prior)
+# Center and scale the input data
+x = aia.get_whitening_layer(transformation="log")(x)
+x_prior = aia.get_whitening_layer(transformation="log")(x_prior)
 x = concatenate([x, x_prior])
 x = Conv2D(4, (1,1), padding='same')(x)
 x = MaxPooling2D(pool_size=(4, 4), strides=4, padding='valid')(x)
@@ -203,16 +194,6 @@ x = Dense(4, activation="relu")(x)
 prediction = Dense(1, activation="linear")(x)
 
 forecaster = Model(inputs=[input_image, input_prior_image, input_side_channel], outputs=[prediction])
-
-tmp1 = np.array([1])
-tmp3 = np.array([3])
-tmp100000 = np.array([100000])
-def custom_loss(y_true, y_pred):
-    return K.tf.pow(K.tf.add(K.tf.multiply(K.tf.abs(y_pred - y_true), tmp100000), 1), tmp3)
-    #return K.tf.log(K.tf.add(K.tf.abs(y_pred - y_true), tmp))
-
-#forecaster.compile(optimizer=adam(lr=0.001, beta_1=0.9, beta_2=0.999, epsilon=1e-08, decay=0.0001), loss="mean_squared_logarithmic_error")
-#forecaster.compile(optimizer=adam(lr=0.001, beta_1=0.9, beta_2=0.999, epsilon=1e-08, decay=0.0001), loss=custom_loss)
 forecaster.compile(optimizer=config["optimizer"], loss=config["loss"])
 
 # Print the netwrok summary information
